+++ changelog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+### 3.1
+- Changed setCooldown-method's return to `this`
+- Added possibility to change help labels (by clearing them or removing specific ones)
+- Added possibility to clear all commands
+- Added possibility to set the embed color for the help command with `CommandSettings#setHelpCommandColor(Color)`
+- Added methods to remove Collections of labels from the commands
+- Added methods to add Collections of labels to the commands
+- Added public Getters for the registered labels, help labels and the boolean `activated`
+- Added first steps for **logging**
+- Changed the split regex for arguments from " " to "\\s+" (it doesn't matter how many spaces there are now)
+- Changed the help command listing type
+- Added some possibly-helpful  methods to class `Command`, such as joining arguments
+- Added channel blacklist
+- Fixed an issue where help labels case sensitivity ignored whether you would activate `labelIgnoreCase`
+- Cleaned some code and fixed some minor bugs
+
 ### 3.0_3
 - Changed parameters of info-method from Guild to Member
 - Added possibility to set command cooldown (either in the constructor of `CommandSettings` or in its setter)